backend/requirements.txt
#--extra-index-url https://pypi.org/simple
#--extra-index-url https://mirrors.aliyun.com/pypi/simple
# FastAPI 框架
fastapi
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
pymilvus==2.6.1
milvus-lite==2.5.1
elasticsearch==9.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1
redis==5.0.1
asyncpg==0.29.0
aioredis==2.0.1

# 向量化和嵌入
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==2.1.1+cpu
# torchvision==0.16.1+cpu
# torchaudio==2.1.1+cpu

# ONNX Runtime for GPU acceleration
# onnxruntime-gpu==1.19.0
# optimum[onnxruntime-gpu]==1.17.1
# onnxruntime
# optimum
optimum[onnxruntime]

# NVIDIA cuDNN for CUDA 12.x
#nvidia-cudnn-cu12==8.9.2.26

# 更新版本以确保兼容性
sentence-transformers==4.1.0
transformers==4.56.1
huggingface_hub==0.34.4
llama-index-embeddings-huggingface==0.6.0
llama-index-vector-stores-milvus==0.9.3
llama-index-vector-stores-elasticsearch==0.6.0

# LLM 集成
openai==1.105.0
langchain==0.3.27
langchain-openai==0.3.32
llama-index==0.13.4

# 文档处理
beautifulsoup4==4.13.5
html2text==2025.4.15

# 文本处理
nltk==3.9.1
spacy==3.7.2

# GraphRAG 依赖
networkx==3.2.1
neo4j==5.15.0
scipy==1.11.4
scikit-learn==1.3.2

# 工具库
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.25.2
aiofiles==23.2.1
psutil==7.1.2
onnx==1.19.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# 日志和监控
loguru==0.7.2
# prometheus-client==0.19.0

# 开发工具
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0

uvicorn[standard]==0.24.0
